tests/e2e/requirements.txt
# Project requirements
# 'pip install -r requirements.txt'

virtualenv
pytest
pyyaml
kiali-client
